requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.3.0
chromadb>=0.4.18
pydantic>=2.5.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
pytesseract>=0.3.10
Pillow>=10.1.0

# UI and visualization
streamlit>=1.29.0
streamlit-chat>=0.1.1

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.1
numpy>=1.26.0
pandas>=2.1.0